Move addToList, removeFromList, and services to transform-edit-base
